uudisreader/requirements.txt
Django==1.6.2
dj-database-url==0.3.0
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
feedparser==5.1.3
numpy==1.7.1
amqp==1.4.4
celery==3.1.10
django-celery==3.1.9
http://math.ut.ee/~ppriit/django-compressor-django-compressor-aba8c5f.tar.gz
django-allauth==0.12.0
selenium==2.41.0
django-voting==0.2